Revert "Define _cubrid.Warning as required by the DB API 2.0"
This definition of Warning redefines the builtin Warning, potentially
causing issues. The builting Warning was not defined when the PEP 249
for DB API 2.0 was created. A note in the DB API:
In a future revision of the DB-API, the base class for Warning will likely
change to the builtin Warning class. At the time of writing of the DB-API
2.0 in 1999, the warning framework in Python did not yet exist.

This reverts commit b90e83c5ae626f43476bcc51213e08b340618bf2.