Implement parameter binding with int, float, date, time, datetime
Sometimes, it is necessary to be possible to bind the parameters as
int, float, date, time, or datetime Python types to the appropriate
CUBRID database types, instead of only providing a string and converting
them. The DBMS will know the intended type and work accordingly.

With these modifications, for example, if an int argument is provided to
bind_param(), CUBRID-cci will take an integer and bind as an integer,
instead of converting to str, passing the str to cubrind and binding as
CHAR and converting back to a number in the database.

This implies fewer type conversions back-and-forth, as an added benefit.